+++ backend/src/main/kotlin/org/n1/av2/script/type/ScriptTypeEntityAndRepo.kt
@@ -16,7 +16,7 @@ data class ScriptType(
     val category: String,
     val size: Int,
     val effects: List<ScriptEffect>,
-    val defaultPrice: BigDecimal?,
+    val defaultPrice: Int?,
 )
 
 data class ScriptEffect(
